requirements.txt
requests
beautifulsoup4
PyYAML
SQLAlchemy
numpy
pandas
scikit-learn
matplotlib
tensorflow==2.15.0
jinja2
schedule